third_party/blink/web_tests/external/wpt/web-animations/responsive/textIndent-expected.txt

This is a testharness.js-based test.
[FAIL] textIndent responsive to inherited textIndent changes
  assert_equals: expected "200px hanging each-line" but got "300px hanging"
Harness: the test ran to completion.
